backend/requirements.txt
google-genai>=1.0.0
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.18.0
fastapi>=0.115.0
uvicorn>=0.30.0
slack-sdk>=3.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
